Add module caching with lua pools and depedency tracking
Loading really large (generated) lua modules turned
out to be really slow, I had to resort to reusing
lua states, which is done by adding lua state pool.
Alas, this means I need to know when a module
has been changed, so I need to know which dependencies
of a module are, so those module can also reloaded.
At least, I managed to avoid doing recusive parsing of modules
to build a dependency tree, by patching lua's require()
so that the dependency are built on the go, as they
are executed.

I don't think this added complexity is really needed,
but on the plus side, this feature is fairly isolated
from the rest of the system, and can be disabled with
a flag.